+++ src/aui-form-field/js/aui-form-field-choice.js
@@ -5,15 +5,26 @@
  */
 
 var CSS_FIELD_CHOICE = A.getClassName('form', 'field', 'choice'),
-    CSS_FIELD_CHOICE_MULTIPLE = A.getClassName('form', 'field', 'choice', 'multiple'),
+    CSS_FIELD_CHOICE_CHECKBOX = A.getClassName('form', 'field', 'choice', 'checkbox'),
+    CSS_FIELD_CHOICE_DROPDOWN = A.getClassName('form', 'field', 'choice', 'dropdown'),
+    CSS_FIELD_CHOICE_DROPDOWN_CONTENT = A.getClassName('form', 'field', 'choice', 'dropdown', 'content'),
+    CSS_FIELD_CHOICE_DROPDOWN_PLACEHOLDER = A.getClassName('form', 'field', 'choice', 'dropdown', 'placeholder'),
+    CSS_FIELD_CHOICE_DROPDOWN_TOGGLE = A.getClassName('form', 'field', 'choice', 'dropdown', 'toggle'),
     CSS_FIELD_CHOICE_OPTION = A.getClassName('form', 'field', 'choice', 'option'),
     CSS_FIELD_CHOICE_OPTION_BUTTON = A.getClassName('form', 'field', 'choice', 'option', 'button'),
     CSS_FIELD_CHOICE_OPTION_OTHER = A.getClassName('form', 'field', 'choice', 'option', 'other'),
     CSS_FIELD_CHOICE_OPTIONS = A.getClassName('form', 'field', 'choice', 'options'),
+    CSS_FIELD_CHOICE_RADIO = A.getClassName('form', 'field', 'choice', 'radio'),
 
     TPL_FIELD_CHOICE_OPTION = '<div class="' + CSS_FIELD_CHOICE_OPTION + ' cleafix">' +
         '<div class="' + CSS_FIELD_CHOICE_OPTION_BUTTON + '"></div>' +
-        '<label>{label}</label></div>';
+        '<label>{label}</label></div>',
+
+    TYPES = {
+        RADIO: 'radio',
+        CHECKBOX: 'checkbox',
+        DROPDOWN: 'dropdown'
+    };
 
 /**
  * A base class for Form Field Choice.
@@ -25,7 +36,11 @@ var CSS_FIELD_CHOICE = A.getClassName('form', 'field', 'choice'),
  * @constructor
  */
 A.FormFieldChoice = A.Base.create('form-field-choice', A.FormField, [], {
-    TPL_FIELD_CONTENT: '<div class="' + CSS_FIELD_CHOICE_OPTIONS + ' form-group"></div>',
+    TPL_FIELD_CONTENT: '<div class="' + CSS_FIELD_CHOICE_OPTIONS + ' form-group"></div>' +
+        '<div class="' + CSS_FIELD_CHOICE_DROPDOWN_CONTENT + '">' +
+        '<div class="' + CSS_FIELD_CHOICE_DROPDOWN_TOGGLE + '">' +
+        '<span class="glyphicon glyphicon-chevron-down"></span></div>' +
+        '<div class="' + CSS_FIELD_CHOICE_DROPDOWN_PLACEHOLDER + '">Select an option</div></div>',
 
     /**
      * Constructor for the `A.FormFieldChoice`. Lifecycle.
@@ -35,9 +50,9 @@ A.FormFieldChoice = A.Base.create('form-field-choice', A.FormField, [], {
      */
     initializer: function() {
         this.after({
-            multipleChange: this._afterMultipleChange,
             optionsChange: this._afterOptionsChange,
-            otherOptionChange: this._afterOtherOptionChange
+            otherOptionChange: this._afterOtherOptionChange,
+            typeChange: this._afterTypeChange
         });
     },
 
@@ -54,21 +69,9 @@ A.FormFieldChoice = A.Base.create('form-field-choice', A.FormField, [], {
 
         content.addClass(CSS_FIELD_CHOICE);
 
-        this._uiSetMultiple(this.get('multiple'));
         this._uiSetOptions(this.get('options'));
         this._uiSetOtherOption(this.get('otherOption'));
-
-
-    },
-
-    /**
-     * Fired after the `multiple` attribute is set.
-     *
-     * @method _afterMultipleChange
-     * @protected
-     */
-    _afterMultipleChange: function() {
-        this._uiSetMultiple(this.get('multiple'));
+        this._uiSetType(this.get('type'));
     },
 
     /**
@@ -92,14 +95,34 @@ A.FormFieldChoice = A.Base.create('form-field-choice', A.FormField, [], {
     },
 
     /**
-     * Updates the ui according to the value of the `multiple` attribute.
+     * Fired after the `type` attribute is set.
      *
-     * @method _uiSetMultiple
-     * @param {Boolean} multiple
+     * @method _afterTypeChange
+     * @param {EventFacade} event
      * @protected
      */
-    _uiSetMultiple: function(multiple) {
-        this.get('content').toggleClass(CSS_FIELD_CHOICE_MULTIPLE, multiple);
+    _afterTypeChange: function(event) {
+        this.get('content').removeClass(this._getClassForType(event.prevVal));
+        this._uiSetType(this.get('type'));
+    },
+
+    /**
+     * Gets the appropriate css class for the given type.
+     *
+     * @method _getClassForType
+     * @param {String} type
+     * @return {String}
+     * @protected
+     */
+    _getClassForType: function(type) {
+        switch (type) {
+            case TYPES.CHECKBOX:
+                return CSS_FIELD_CHOICE_CHECKBOX;
+            case TYPES.RADIO:
+                return CSS_FIELD_CHOICE_RADIO;
+            case TYPES.DROPDOWN:
+                return CSS_FIELD_CHOICE_DROPDOWN;
+        }
     },
 
     /**
@@ -149,6 +172,17 @@ A.FormFieldChoice = A.Base.create('form-field-choice', A.FormField, [], {
                 optionNode.remove(true);
             }
         }
+    },
+
+    /**
+     * Updates the ui according to the value of the `type` attribute.
+     *
+     * @method _uiSetType
+     * @param {String} type
+     * @protected
+     */
+    _uiSetType: function(type) {
+        this.get('content').addClass(this._getClassForType(type));
     }
 }, {
     /**
@@ -206,6 +240,25 @@ A.FormFieldChoice = A.Base.create('form-field-choice', A.FormField, [], {
         required: {
             validator: A.Lang.isBoolean,
             value: false
+        },
+
+        /**
+         * The type of the choice field. Can be any of the values listed in
+         * the `A.FormFieldChoice.TYPES` constant (radio, checkbox or dropdown);
+         * @type {Object}
+         */
+        type: {
+            validator: A.Lang.isString,
+            value: TYPES.RADIO
         }
-    }
+    },
+
+    /**
+     * Static property which contains all the valid `A.FormFieldChoice` types.
+     *
+     * @property TYPES
+     * @type Object
+     * @static
+     */
+    TYPES: TYPES
 });