+++ src/aui-form-field/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-1785](https://issues.liferay.com/browse/AUI-1785) Merge list and choice fields into a single field type
 * [AUI-1794](https://issues.liferay.com/browse/AUI-1794) Refactor the rendering logic of each field into a render function
 * [AUI-1750](https://issues.liferay.com/browse/AUI-1750) Create time field
 * [AUI-1784](https://issues.liferay.com/browse/AUI-1784) Convert existing form builder fields to separate rendering from edition logic